Move /createTournament request response to controller
